Adds map_shm to open and mmap a region of POSIX shared memory.
The map_shm function takes a name variable instead of an fd or file.
First, map_shm creates and/or attaches to a region of shared memory, which
returns an fd.  It then calls the usual mmap function with the fd.
Finally, it adds an unlink function to the returned buffer.  The unlink
function marks the shared memory region for deletion, once all processes
have detached using unmap (which can be called explicitly or implicitly
during destruction as usual).

All other functionality of the returned buffer remains the same.